refactor: extract logo button from header and create individual component
