+++ src/components/LogoButton/LogoButton.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link';
 
 const LogoButton = () => (
   <Link href="/">
-    <Image width={140} height={32} src="/logo.svg" alt="chillka" />
+    <Image width={140} height={32} src="/logo.svg" alt="chillka logo" />
   </Link>
 );
