+++ src/components/LogoButton/index.ts
@@ -1,0 +1,3 @@
+import LogoButton from './LogoButton';
+
+export default LogoButton;
